+++ src/App.jsx
@@ -9,6 +9,7 @@ import RecipesPage from './pages/RecipesPage';
 import LoginPage from './pages/LoginPage';
 import SignupPage from './pages/SignupPage';
 import AccountPage from './pages/AccountPage';
+import PrivacyPage from './pages/PrivacyPage';
 import { isOcrEnabled } from './utils/backendConfig';
 
 function App() {
@@ -25,6 +26,7 @@ function App() {
         <Route path="/recipes" element={<RecipesPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/signup" element={<SignupPage />} />
+        <Route path="/privacy" element={<PrivacyPage />} />
         <Route
           path="/account"
           element={